dotfiles on GitHub Actions

+++ vscode/restore.sh
@@ -3,15 +3,18 @@
 SCRIPT_DIR=$(cd $(dirname $0) && pwd)
 VSCODE_SETTING_DIR=~/Library/Application\ Support/Code/User
 
-rm "$VSCODE_SETTING_DIR/settings.json"
-ln -s "$SCRIPT_DIR/settings.json" "${VSCODE_SETTING_DIR}/settings.json"
+if [ -e "$VSCODE_SETTING_DIR/settings.json" ]; then
+  rm "$VSCODE_SETTING_DIR/settings.json"
+  ln -s "$SCRIPT_DIR/settings.json" "${VSCODE_SETTING_DIR}/settings.json"
+fi
 
-rm "$VSCODE_SETTING_DIR/keybindings.json"
-ln -s "$SCRIPT_DIR/keybindings.json" "${VSCODE_SETTING_DIR}/keybindings.json"
+if [ -e "$VSCODE_SETTING_DIR/keybindings.json" ]; then
+  rm "$VSCODE_SETTING_DIR/keybindings.json"
+  ln -s "$SCRIPT_DIR/keybindings.json" "${VSCODE_SETTING_DIR}/keybindings.json"
+fi
 
 cat ./extensions | while read line
 do
   code --install-extension $line
 done
 code --list-extensions > extensions
-
